requirements.txt
openai
python-dotenv
black
ruff
jupyter
scikit-learn
pandas
